+++ frontend/src/metabase/dashboard/actions/index.ts
@@ -10,3 +10,4 @@ export * from "./sharing";
 export * from "./ui";
 export * from "./actions";
 export * from "./tabs";
+export * from "./trash";
